MR2PhaseApriori: noreupload feature added

+++ java/Eclipse-workspace/MR-2Phase-Apriori/src/data.sh
@@ -7,9 +7,16 @@ cd $abshere
 
 folder=`basename $1`
 
-ssh -n hadoop1 rm -rf /tmp/$folder
-scp -r $1 hadoop1:/tmp/
-ssh -n hadoop1 ./hadoop-2.2.0/bin/hdfs dfs -rm -r -f /input-test /output-test-1stphase /output-test
+ssh -n hadoop1 ./hadoop-2.2.0/bin/hdfs dfs -rm -r -f /output-test-1stphase /output-test
 
-ssh -n hadoop1 ./hadoop-2.2.0/bin/hdfs dfs -mkdir /input-test
-ssh -n hadoop1 ./hadoop-2.2.0/bin/hdfs dfs -copyFromLocal /tmp/$folder/* /input-test/
+if [ "$2" = "noreupload" ]; then
+  echo Skip data reupload, use previously uploaded input
+else
+  ssh -n hadoop1 rm -rf /tmp/$folder
+  scp -r $1 hadoop1:/tmp/
+  ssh -n hadoop1 ./hadoop-2.2.0/bin/hdfs dfs -rm -r -f /input-test
+
+  ssh -n hadoop1 ./hadoop-2.2.0/bin/hdfs dfs -mkdir /input-test
+  ssh -n hadoop1 ./hadoop-2.2.0/bin/hdfs dfs -copyFromLocal /tmp/$folder/* /input-test/
+
+fi
